refactor inference
